Istio: Update to 1.6.14
On this version the sidecar configuration is changed to not use TPROXY
redirection, as support for it has been broken in Istio upstream.

Test helper `ExpectCiliumOperatorReady()` is adjusted to work in
single-node clusters, where it is expected that only one of the Cilium
Operator instances comes up.

Istio CI test is now unquarantined on K8s 1.19.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>